Updating the new challange view

+++ server/public/scripts/services/challenges.service.js
@@ -1,11 +1,11 @@
 myApp.factory('ChallengesService', function($http, $location) {
   console.log('ChallengesService Loaded');
 
- var challengesService = {
+  var challengesService = {
 
-   challenges: [],
+    challenges: [],
 
-   getChallenges: function() {
+    getChallenges: function() {
       console.log('get challenges');
       $http.get('/challenges').then(function(response) {
         var startDate = Date.parse(response.data[0].start_date)
@@ -16,18 +16,7 @@ myApp.factory('ChallengesService', function($http, $location) {
       });
     }, // end getChallenges
 
-
-// Ale start
-    // acceptStudentChallenge: function() {
-    //   console.log('accept student challenges');
-    //   $http.get('/challenges/:studentId').then(function(response){
-    //     console.log(response);
-    //     challengesService.challenges = response.data;
-    //   });
-    // }, // end acceptStudentChallenge
-// Ale end
-
-   addChallenge: function(newChallenge) {
+    addChallenge: function(newChallenge) {
       console.log(newChallenge);
       $http.post('/challenges', newChallenge).then(function(response) {
         console.log(response);
@@ -35,18 +24,7 @@ myApp.factory('ChallengesService', function($http, $location) {
       });
     }, // end addChallenge
 
-// // Ale start
-//     addChallengeToStudent: function(newStudent) {
-//        console.log(newStudent);
-//        $http.post('/challenges/addStudent', newStudent).then(function(response) {
-//          console.log(response);
-//          challengesService.acceptStudentChallenge();
-//        });
-//      }, // end addChallenge
-//      // Ale end
-
-
-   updateChallenge: function(challenge) {
+    updateChallenge: function(challenge) {
       console.log('update challenge');
       console.log(challenge);
       $http.put('/challenges', challenge).then(function(response) {
@@ -56,28 +34,38 @@ myApp.factory('ChallengesService', function($http, $location) {
     }, // end updateChallenge
 
 
-   deleteChallenge: function(challenge) {
+    deleteChallenge: function(challenge) {
       console.log('deleteChallenge');
       console.log(challenge);
-      $http.delete('/challenges/' + challenge.id).then(function(response) {
-        console.log(response);
-        challengesService.getChallenges();
+      swal({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        type: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#E3968B',
+        cancelButtonColor: '#839EAC',
+        confirmButtonText: 'Yes, delete it!'
+      }).then(function () {
+        $http.delete('/challenges/' + challenge.id).then(function(response) {
+          console.log(response);
+          challengesService.getChallenges();
+        });
       });
     }, // end deleteChallenge
 
     acceptChallenge : function(challengeId) {
       $http.post('/students/' + challengeId).then(function(response) {
-      console.log(response);
-      swal(
-        'Good job!',
-        'You selected a challange!',
-        'success'
-      );
-    });
-  } // end acceptChallenge
+        console.log(response);
+        swal(
+          'Good job!',
+          'You selected a challange!',
+          'success'
+        );
+      });
+    } // end acceptChallenge
 
   };
 
- return challengesService;
+  return challengesService;
 
 }); // end challenges service
